refactor: restructure init.lua and modularize keymaps and defaults
- Moved keymap setup to a dedicated module for better organization and maintainability.
- Introduced a new defaults module to encapsulate default settings and configurations.
- Added a lazy loading setup for plugins to improve startup time.
- Removed deprecated autoformat configurations and unnecessary comments for clarity.
- Cleaned up the init.lua file by delegating responsibilities to respective modules.
