+++ lua/phoenix/plugins/nvim-colorizer.lua
@@ -1,0 +1,14 @@
+  return {
+    'norcalli/nvim-colorizer.lua',
+    config = function()
+      require('colorizer').setup {
+        'css',
+        'javascript',
+        'json',
+        'typescriptreact',
+        'typescript',
+        'lua',
+      }
+    end,
+  }
+
