+++ lua/phoenix/plugins/telescope.lua
@@ -1,0 +1,85 @@
+local map = require 'phoenix.utils.keymap'
+
+return { -- Fuzzy Finder (files, lsp, etc)
+  'nvim-telescope/telescope.nvim',
+  event = 'VimEnter',
+  dependencies = {
+    'nvim-lua/plenary.nvim',
+    { -- If encountering errors, see telescope-fzf-native README for installation instructions
+      'nvim-telescope/telescope-fzf-native.nvim',
+
+      -- `build` is used to run some command when the plugin is installed/updated.
+      -- This is only run then, not every time Neovim starts up.
+      build = 'make',
+
+      -- `cond` is a condition used to determine whether this plugin should be
+      -- installed and loaded.
+      cond = function()
+        return vim.fn.executable 'make' == 1
+      end,
+    },
+    { 'Marskey/telescope-sg' },
+    { 'nvim-telescope/telescope-ui-select.nvim' },
+
+    -- Useful for getting pretty icons, but requires a Nerd Font.
+    { 'nvim-tree/nvim-web-devicons',            enabled = vim.g.have_nerd_font },
+  },
+  config = function()
+    require('telescope').setup {
+      defaults = {
+        mappings = {
+          i = {
+            ['<C-u>'] = false,
+            ['<C-d>'] = false,
+          },
+        },
+        path_display = {
+          shorten = { len = 1, exclude = { -2, -1 } },
+        },
+        dynamic_preview_title = true,
+      },
+      extensions = {
+        fzf = {
+          override_generic_sorter = true, -- override the generic sorter
+          override_file_sorter = true,    -- override the file sorter
+          case_mode = 'smart_case',       -- or "ignore_case" or "respect_case"
+        },
+        ['ui-select'] = {
+          require('telescope.themes').get_dropdown(),
+        },
+      },
+    }
+
+    -- Enable Telescope extensions if they are installed
+    pcall(require('telescope').load_extension, 'fzf')
+    pcall(require('telescope').load_extension, 'ui-select')
+    pcall(require('telescope').load_extension 'ast_grep')
+
+    -- See `:help telescope.builtin`
+    local builtin = require 'telescope.builtin'
+    map('n', '<leader>sf', builtin.find_files, { desc = '[S]earch [F]iles' })
+    map('n', '<leader>sw', builtin.grep_string, { desc = '[S]earch current [W]ord' })
+    map('n', '<leader>sg', builtin.live_grep, { desc = '[S]earch by [G]rep' })
+    map('n', '<leader>sd', builtin.diagnostics, { desc = '[S]earch [D]iagnostics' })
+    map('n', '<leader>sr', builtin.resume, { desc = '[S]earch [R]esume' })
+    map('n', '<leader>?', builtin.oldfiles, { desc = 'Search Recent Files' })
+    map('n', '<leader>b', builtin.buffers, { desc = '[ ] Find existing buffers' })
+    map('n', '<leader>gs', require('telescope.builtin').git_status, { desc = '[G]it [S]tatus' })
+
+    -- Slightly advanced example of overriding default behavior and theme
+
+    -- It's also possible to pass additional configuration options.
+    --  See `:help telescope.builtin.live_grep()` for information about particular keys
+    map('n', '<leader>s/', function()
+      builtin.live_grep {
+        grep_open_files = true,
+        prompt_title = 'Live Grep in Open Files',
+      }
+    end, { desc = '[S]earch [/] in Open Files' })
+
+    -- Shortcut for searching your Neovim configuration files
+    map('n', '<leader>sn', function()
+      builtin.find_files { cwd = vim.fn.stdpath 'config' }
+    end, { desc = '[S]earch [N]eovim files' })
+  end,
+}
